feat: domain을 entity로 변환
