+++ src/main/java/roomescape/controller/api/AuthController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.config.annotation.AuthMember;
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 import roomescape.dto.request.LoginRequest;
 import roomescape.dto.response.LoginResponse;
 import roomescape.service.AuthService;
