+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -2,7 +2,7 @@ package roomescape.dao;
 
 import java.time.LocalDate;
 import java.util.List;
-import roomescape.domain.Reservation;
+import roomescape.entity.Reservation;
 
 public interface ReservationDao {
