+++ src/main/java/roomescape/dao/ThemeDao.java
@@ -2,7 +2,7 @@ package roomescape.dao;
 
 import java.time.LocalDate;
 import java.util.List;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 
 public interface ThemeDao {
