+++ src/main/java/roomescape/dao/jdbc/JdbcMemberDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.dao.MemberDao;
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 import roomescape.exception.custom.NotFoundException;
 
 @Repository
