--- src/main/java/roomescape/domain/Member.java
@@ -1,36 +1,0 @@
-package roomescape.domain;
-
-public class Member {
-
-    private final Long id;
-    private final String name;
-    private final String email;
-    private final String password;
-
-    public Member(Long id, String name, String email, String password) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-        this.password = password;
-    }
-
-    public Member(String name, String email, String password) {
-        this(null, name, email, password);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
