--- src/main/java/roomescape/domain/Theme.java
@@ -1,36 +1,0 @@
-package roomescape.domain;
-
-public class Theme {
-
-    private final Long id;
-    private final String name;
-    private final String description;
-    private final String thumbnail;
-
-    public Theme(Long id, String name, String description, String thumbnail) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.thumbnail = thumbnail;
-    }
-
-    public Theme(String name, String description, String thumbnail) {
-        this(null, name, description, thumbnail);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getThumbnail() {
-        return thumbnail;
-    }
-}
