+++ src/main/java/roomescape/dto/response/MemberResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto.response;
 
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 
 public record MemberResponse(
     Long id,
