+++ src/main/java/roomescape/provider/JwtTokenProvider.java
@@ -4,7 +4,7 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 
 @Component
 public class JwtTokenProvider {
