+++ src/main/java/roomescape/service/TimeService.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 import roomescape.dto.request.TimeRequest;
 import roomescape.exception.custom.DuplicatedException;
