+++ src/test/java/roomescape/dao/jdbc/JdbcMemberDaoTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 import roomescape.exception.custom.NotFoundException;
 
 @JdbcTest
