+++ src/test/java/roomescape/dao/jdbc/JdbcThemeDaoTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.domain.Theme;
+import roomescape.entity.Theme;
 import roomescape.exception.custom.NotFoundException;
 
 @JdbcTest
