+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -17,10 +17,10 @@ import roomescape.dao.fake.FakeMemberDao;
 import roomescape.dao.fake.FakeReservationDao;
 import roomescape.dao.fake.FakeReservationTimeDao;
 import roomescape.dao.fake.FakeThemeDao;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
+import roomescape.entity.Member;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
+import roomescape.entity.Theme;
 import roomescape.dto.request.ReservationRequest;
 import roomescape.exception.custom.DuplicatedException;
 import roomescape.exception.custom.InvalidInputException;
